Added the ahk script. It stays open upon being executed and has a hotkey that can be used to execute the main.ps1 file. The hotkey is shift+F1. A console will open up that prints the output of SetDefaultPrinter and asks the user for confirmation before closing. Also updated the readme accordingly with some brief instructions.
<Reflection>

It contains some commented out code that wasn't printing out text like I expected it to. In hindsight, I believe I was ordering the -Command and -NoExit errors erroneously. Since it's commented out, I'm going to leave the code-lines as a record for the type of error my mind generated and hopefully learn more from it going forward.

Hint for future Caleb: if the written patterns you're using aren't resulting in the expected output, perhaps an alternative pattern for typing powershell lines will more correctly align with the underlying design philosophy of the language/software. Relevant example: Remember that parameter names are prepended by a '-'. Any non-dash-prepended text that follows will be the value assigned to that parameter. It's not that order necessarily matters (Sometimes a list of parameters are commutative. Is associativity ever relevant in evaluation order? --> Perhaps if the left operand results in a returned object type that the right operand doesn't accept.), but that if a parameter_name is followed by a parameter_value, the value will necessarily be bound to that parameter.
https://learn.microsoft.com/en-us/powershell/module/microsoft.powershell.core/about/about_parameters?view=powershell-7.4
It's maybe redundant and self-evident, but there's plenty of value in returning to simplest logics when one fails to overcome their misunderstanding because they misunderstand the space in the first place. Note that I only managed to correctly think through the problem by simply returning to it after moving onto a different, and ultimately working, solution. It would appear that perspective shifts, even random path-finding, is rather effecient.

</Reflection>